% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESliceSampling.R
\name{ESlice}
\alias{ESlice}
\title{Eliptical slice sampling}
\usage{
ESlice(f_cur, OdeTraj, FTs, state, init, t_correct, lambda = 10, reps = 1L,
  gridsize = 100L, funname = "standard")
}
\arguments{
\item{f_cur}{Current trajectory}

\item{OdeTraj}{Ode trajectory of the given parameter}

\item{state}{Initial state}

\item{t_correct}{time correction for coalescent times}

\item{lambda}{scale parameter for effective population}

\item{reps}{Repeat times}
}

