% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESliceSampling.R
\name{SIR_LNA_ESlice_MCMC}
\alias{SIR_LNA_ESlice_MCMC}
\title{theta1
theta2
Trajectory matrix: in log-scale
state Initial state in 2D (X,Y)
OdeTraj_thin: ode trajectory with high resoluton, used to generate the filter
OdeTraj_coarse: ode trajectory with low resultion, used to generate the path
LatentTraj: the latent SIR 3d trajectory, has the same resolution as OdeTraj_coarse
logMultNorm: loglikelihood of Ode trajectory}
\usage{
SIR_LNA_ESlice_MCMC(state, Init, times, gridsize = 100, niter = 1000,
  burn = 500, thin = 5)
}
\description{
Infectious rate and recovery rate: theta1, theta2 
Scale parameter: lambda
}
\details{
prior
s1:  s1 = theta1  log(1/s1) ~ gamma(50,5)
propose function random walk Uniform[-0.5,0.5] on log space

s2: theta2/theta1   proposal function: random walk Uniform[-0.5,0.5]
Normal(20000,2000)

lambda log(lambda) ~ gamma(16,4) propose random walk uniform [-0.5,0.5] on logspace
}

