% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_LNA_ESlice_MCMC4.R
\name{SIR_LNA_ESlice_MCMC4}
\alias{SIR_LNA_ESlice_MCMC4}
\title{theta1
theta2
Trajectory matrix: in log-scale
state Initial state in 2D (X,Y)
OdeTraj_thin: ode trajectory with high resoluton, used to generate the filter
OdeTraj_coarse: ode trajectory with low resultion, used to generate the path
LatentTraj: the latent SIR 3d trajectory, has the same resolution as OdeTraj_coarse
logMultNorm: loglikelihood of Ode trajectory}
\usage{
SIR_LNA_ESlice_MCMC4(Init, times, t_correct, N, gridsize = 50, niter = 1000,
  burn = 500, thin = 5, a1 = 10, a2 = 20, c1 = -2.3, d1 = 200,
  d2 = 40)
}
\description{
Infectious rate and recovery rate: theta1, theta2
Scale parameter: lambda
}
\details{
alpha1: log(alpha1) ~ gamma(60,a1)
alpha2: log(alpha2) ~ gamma(60,a2)
prior
s1:  s1 = theta1*N/theta2  U[1,10]
propose function random walk Uniform[-0.5,0.5] on log space

s2: theta2 log(s2) ~ norm(c1,0.4)

lambda log(lambda) ~ gamma(d1,d2) propose random walk uniform [-0.5,0.5] on logspace

log(p1/p3) ~ gamma(60,10)
log(p2/p3) ~ gamma(60,20)
}

